+++ docs/changes.rst
@@ -5,6 +5,7 @@ Pending
 -------
 
 * Added hook to RedirectsPanel for subclass customization.
+* Added feature to sanitize sensitive data in the Request Panel.
 
 5.1.0 (2025-03-20)
 ------------------